+++ R/00-Namespace.R
@@ -5,21 +5,21 @@
 #' @importFrom rcarbon calibrate sampleDates
 #' @importFrom DataTools cutAllLongStrings downUploadButtonUI downUploadButtonServer
 #' importDataUI importDataServer
-#' @importFrom dplyr group_by_at select summarise left_join
+#' @importFrom dplyr arrange distinct filter group_by group_by_at left_join select summarise ungroup
 #' @importFrom DT datatable renderDataTable dataTableOutput
 #' @importFrom geometry convhulln inhulln
 #' @importFrom ggplot2 ggplot theme theme_light coord_cartesian geom_point theme_light theme labs
 #'  geom_errorbar aes_ element_blank element_text position_dodge aes geom_boxplot xlab
 #' @importFrom graphics axis filled.contour hist plot points title boxplot par image lines
 #'  polygon text .filled.contour Axis box layout lcm plot.new plot.window rect legend strwidth text
-#' @importFrom grDevices colorRampPalette chull png pdf tiff svg dev.off recordPlot replayPlot
-#'  cm.colors pdfFonts
+#' @importFrom grDevices cm.colors colorRampPalette col2rgb chull dev.off png pdf pdfFonts
+#'  recordPlot replayPlot rgb svg tiff
 #' @importFrom jsonlite fromJSON toJSON
 #' @importFrom ks kde Hpi.diag Hpi
-#' @importFrom leaflet addAwesomeMarkers addCircleMarkers addControl addLegend addMarkers
-#'  addPopups addProviderTiles addRectangles addScaleBar addTiles awesomeIcons
-#'  clearMarkerClusters clearGroup clearPopups colorFactor colorNumeric fitBounds
-#'  leaflet leafletProxy leafletOutput markerClusterOptions popupOptions
+#' @importFrom leaflet addAwesomeMarkers addCircleMarkers addControl addLayersControl addLegend
+#'  addMarkers addPopups addProviderTiles addRectangles addScaleBar addTiles awesomeIcons
+#'  clearMarkerClusters clearGroup clearPopups colorFactor colorNumeric fitBounds icons
+#'  layersControlOptions leaflet leafletProxy leafletOutput markerClusterOptions popupOptions
 #'  removeControl removeScaleBar removeShape renderLeaflet scaleBarOptions setMaxBounds setView
 #' @importFrom magrittr "%>%"
 #' @importFrom maps map map.scale
@@ -35,7 +35,7 @@
 #' @importFrom shinyjs alert runjs useShinyjs hide show
 #' @importFrom shinyMatrix matrixInput updateMatrixInput
 #' @importFrom shinyWidgets pickerInput updatePickerInput
-#' @importFrom stats aggregate as.formula cor cov density dnorm dunif kmeans
+#' @importFrom stats aggregate as.formula cor cov density dist dnorm dunif kmeans
 #'  mahalanobis median model.matrix na.exclude na.omit pchisq  pf pnorm predict
 #'  qnorm quantile rbeta residuals rgamma rnorm runif sd var
 #' @importFrom sp point.in.polygon plot SpatialPoints