+++ R/03-modelResults2D.R
@@ -265,7 +265,7 @@ modelResults2DUI <- function(id, title = "", asFruitsTab = FALSE){
                       selected = "RdYlGn"),
           ns = ns),
         selectInput(inputId = ns("pointShape"), label = "Shape of location marks",
-                    choices = 0:25, selected = 4),
+                    choices = pchChoices(), selected = 4),
         ns = ns),
         checkboxInput(inputId = ns("grid"),
                       label = "Show map grid",