+++ R/03-modelResultsSpread.R
@@ -328,7 +328,7 @@ modelResultsSpreadUI <- function(id, title = ""){
                           selected = "RdYlGn"),
               ns = ns),
             selectInput(inputId = ns("pointShape"), label = "Shape of location marks",
-                        choices = 0:25, selected = 4),
+                        choices = pchChoices(), selected = 4),
             ns = ns),
           checkboxInput(inputId = ns("interior"),
                         label = "Apply convex hull",